new changes

+++ src/app/pages/edit-chat-page/chat-data-edit/controls-edit/controls-edit.component.ts
@@ -53,6 +53,7 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
   outletControl: AnswerInterface | null = null;
   outletRequired = false;
   isPersonIdentifying = false;
+  dropElSelectedData: string[] = [];
 
   constructor(
     private route: ActivatedRoute,
@@ -106,6 +107,7 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
           }
 
           this.currentId = current_data?.id;
+          this.generateDropValue();
           this.controls = chat_data;
 
           const input_control = this.controls.find(val => val.control_type === 'input');
@@ -113,10 +115,29 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
           const el_controls = chat_data.filter(el => el?.control_type !== 'button-next' && el?.control_type !== 'dropdown-item')
           const el_drop_ctrl = chat_data.find (el => el?.control_type === 'dropdown-item');
 
+          const full_data = this.customerService.getProgress();
           // set controls
           for (let el of [...el_controls, el_drop_ctrl]) {
             if(el) {
-              this.form.addControl(el.id.toString(), this.getInputValidation(el?.control_validation))
+              this.form.addControl(el.id.toString(), this.getInputValidation(el?.control_validation));
+
+              if(full_data && el.id) {
+                const found = full_data.find(val => val.answer_id == el?.id);
+
+                console.log(1, found)
+                if(found?.clear_text) {
+                  this.form.patchValue({[el.id.toString()]: found?.clear_text});
+
+                  if(el?.control_text === 'Geburtsdatum') {
+                    this.form.patchValue({[el.id.toString()]: moment(found?.clear_text, "DD.MM.YYYY").toDate()});
+                  }
+                  if(el?.control_type === 'dropdown-item') {
+                    // console.log(3, found, el);
+                    this.form.patchValue({[el.id.toString()]: found});
+                  }
+                }
+
+              }
             }
           }
         }
@@ -129,7 +150,7 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
       return new FormControl('', [Validators.required, Validators.minLength(3)])
     }
     if(type === 'email') {
-      return new FormControl('', [Validators.required, Validators.pattern("^[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,4}$")])
+      return new FormControl('', [Validators.required, Validators.pattern("^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,4}$")])
     }
     if(type === 'phone') {
       return new FormControl('+41',[Validators.required, Validators.pattern(/(\b(0041|0)|\B\+41)(\s?\(0\))?(\s)?[1-9]{2}(\s)?[0-9]{3}(\s)?[0-9]{2}(\s)?[0-9]{2}\b/)]);
@@ -149,8 +170,20 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
     await this.onSelect(e.value, true);
   }
 
-  async onOutlet (control: AnswerInterface) {
-    await this.customerService.removeUserProgress(control?.group_id);
+  generateDropValue() {
+    this.dropElSelectedData = [];
+    this.customerService.getProgress().map(val => {
+      if(val.id === this.currentId) {
+        this.dropElSelectedData.push(String(val.id + ' ' + val.answer_id));
+      }
+    })
+  }
+
+  async onOutlet (control: AnswerInterface, isMultiple?: boolean) {
+    if(!isMultiple) {
+      await this.customerService.removeUserProgress(control?.group_id);
+    }
+
     this.outletControl = control;
 
     const next: any = control?.next_step?.data;
@@ -161,6 +194,7 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
       next_id: next?.id ? next?.id : 0,
       type: 'user',
       text: control.user_message,
+      clear_text: control?.user_message_clear ? control?.user_message_clear : control.user_message,
       text_pdf: control.user_pdf_message,
       // @ts-ignore
       step: control.step,
@@ -202,6 +236,8 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
         await this.customerService.removeUserProgressByStep(control.remove_next_on_edit);
       }
     }
+
+    setTimeout(() => this.generateDropValue());
   }
 
 
@@ -236,6 +272,7 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
       next_id: next?.id ? next.id : 0,
       type: 'user-q',
       text: control.user_message,
+      clear_text: control?.user_message_clear ? control?.user_message_clear : control.user_message,
       text_pdf: control.user_pdf_message,
       // @ts-ignore
       step: control.step,
@@ -288,7 +325,7 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
 
         if(answer?.control_type === 'input' || answer?.control_type === 'input-autocomplete') {
           if(answer?.control_text === "Geburtsdatum") {
-            value.value = moment(value.value).format('DD.MM.YYYY')
+            value.value = moment(value.value).format('DD.MM.YYYY');
           }
 
           if(answer.bot_message) {
@@ -296,11 +333,16 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
           }
           if(answer.user_message) {
             answer.user_message = stringReplace(answer.user_message, value?.value);
+            answer.user_message_clear = value?.value;
             answer.user_pdf_message = stringReplace(answer.user_pdf_message, value?.value);
           }
 
           await this.onSelect(answer);
         } else {
+          if(answer?.control_type === 'dropdown-item' && value?.value) {
+            answer.user_message_clear = value?.value?.id;
+            answer = value?.value;
+          }
           await this.onSelect(answer);
         }
       }
@@ -321,6 +363,7 @@ export class ControlsEditComponent implements OnInit, OnDestroy {
       next_id: next?.id ? next?.id : 0,
       type: 'user',
       text: control.user_message,
+      clear_text: control?.user_message_clear ? control?.user_message_clear : control.user_message,
       text_pdf: control.user_pdf_message,
       // @ts-ignore
       step: control?.step,
